Added intnral funcationality to allow for calculation of profits, different conditions, assignment of role identifiers within groups
I added several functions to the player class so that the cost to buyers,
total seller price, number of sales made, and profit are all calculated at
the end of each round.  Individuals are assigned to roles and for each
individual sharing a role within a group they are randomly assigned either
a value of 1 or 2 in order to identify the particular player within the
round.

The game now lets you specify the number of practice rounds and the number
of games plus the total number of non-practice round games.