The basic functionality of the program exists where players can go through all the screens with one hiccup and play the whole game for an arbitrary number of rounds.
Currently the buyer wait page where individuals can click on the buttons
is not working properly.  It is no longer showing the number of clicks to
individuals and these pages to not automatically push the the round
summary once all individuals have made it through the seller selection
process.

Other than than individuals are shown a very simple summary of what
happens in each round with the prices from each seller.

Many additional formatting tweaks will be required to make the output as
desired but all of the variables needed should be calculated and easy to
pass to the forms.