Changes were made to the display of output and additional work was started on the questions asked for each participant.
The round summary and the seller selection pages now matche the zTree
programs much more closely with short formulas to explain how we arrived
at the round profits for an individual player.

As for the questions to be asked to each participant we still need to work
on the mechanism for sending messages to players that initially select the
wrong response.  Other than that the questions and response options match
the zTree experiment.